+++ src/test/java/org/folio/search/controller/IndexingInstanceCallNumberIT.java
@@ -81,8 +81,10 @@ class IndexingInstanceCallNumberIT extends BaseIntegrationTest {
     assertThat(instances)
       .hasSize(1)
       .allSatisfy(map -> assertThat(map).containsEntry("shared", false))
-      .allSatisfy(map -> assertThat(map).containsEntry("tenantId", TENANT_ID))
-      .allSatisfy(map -> assertThat(map).containsEntry("count", 2));
+      .allSatisfy(map -> assertThat(map).containsEntry("tenantId", TENANT_ID));
+    @SuppressWarnings("unchecked")
+    var ids = (List<String>) instances.get(0).get("instanceId");
+    assertThat(ids).containsExactlyInAnyOrder(INSTANCE_ID_1, INSTANCE_ID_2);
   }
 
   @Test